Refactor for src and extract db logic
